requirements.txt
web3>=6.0.0
py-clob-client>=0.1.0
python-dotenv>=1.0.0
